♻️ separate type definition from component
